Extract common styles for Success and Fail forms
