+++ src/components/irregularity/helpers/Success.tsx
@@ -9,7 +9,7 @@ import CheckCircleOutlinedIcon from '@mui/icons-material/CheckCircleOutlined'
 import theme from 'common/theme'
 import { routes } from 'common/routes'
 
-import { Root, SuccessButton, SuccessButtonsWrapper, SuccessIcon } from './Success.styled'
+import { Root, ActionButton, ButtonsWrapper, Icon } from './Irregularity.styled'
 
 export default function Success() {
   const { t } = useTranslation('irregularity')
@@ -23,9 +23,9 @@ export default function Success() {
   return (
     <ThemeProvider theme={theme}>
       <Root container>
-        <SuccessIcon item xs={12}>
+        <Icon item xs={12} color="#4BD12A">
           <CheckCircleOutlinedIcon />
-        </SuccessIcon>
+        </Icon>
         <Grid item xs={12}>
           <Typography variant="h5">{t('steps.success.title')}</Typography>
         </Grid>
@@ -34,20 +34,20 @@ export default function Success() {
             {t('steps.success.subtitle')}
           </Typography>
         </Grid>
-        <SuccessButtonsWrapper container>
+        <ButtonsWrapper container>
           <Grid item xs={12} sm={6}>
             <Typography variant="body1">{t('steps.success.label-campaigns')}</Typography>
-            <SuccessButton href={routes.campaigns.index} variant="outlined">
+            <ActionButton href={routes.campaigns.index} variant="outlined">
               {t('cta.campaigns')}
-            </SuccessButton>
+            </ActionButton>
           </Grid>
           <Grid item xs={12} sm={6}>
             <Typography variant="body1">{t('steps.success.label-project')}</Typography>
-            <SuccessButton href={routes.aboutProject} variant="outlined">
+            <ActionButton href={routes.aboutProject} variant="outlined">
               {t('cta.project')}
-            </SuccessButton>
+            </ActionButton>
           </Grid>
-        </SuccessButtonsWrapper>
+        </ButtonsWrapper>
       </Root>
     </ThemeProvider>
   )
